regreso

+++ [EDD]Proyecto_2/src/Entidad/Cliente.java
@@ -4,13 +4,13 @@ import java.math.BigInteger;
 
 public class Cliente implements Comparable<Cliente> {
 
-    private BigInteger dpi;
+    private int dpi;
     private String nombre;
     private String apellido;
     private String genero;
     private String direccion;
 
-    public Cliente(BigInteger dpi, String nombre, String apellido, String genero, String direccion) {
+    public Cliente(int dpi, String nombre, String apellido, String genero, String direccion) {
         this.dpi = dpi;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -18,11 +18,11 @@ public class Cliente implements Comparable<Cliente> {
         this.direccion = direccion;
     }
 
-    public BigInteger getDpi() {
+    public int getDpi() {
         return dpi;
     }
 
-    public void setDpi(BigInteger dpi) {
+    public void setDpi(int dpi) {
         this.dpi = dpi;
     }
 
@@ -68,7 +68,7 @@ public class Cliente implements Comparable<Cliente> {
     public int compareTo(Cliente o) {
         if (this.dpi == o.getDpi()) {
             return 0;
-        } else if (this.dpi.intValue() > o.getDpi().intValue()) {
+        } else if (this.dpi > o.getDpi()) {
             return 1;
         } else {
             return -1;
